+++ Tests/BugsnagTests/BugsnagClientMirrorTest.m
@@ -76,6 +76,8 @@
             @"orientationDidChange: v24@0:8@16",
             @"pluginClient @16@0:8",
             @"populateEventData: v24@0:8@16",
+            @"readyForInternalCalls B16@0:8",
+            @"readyForInternalCalls c16@0:8",
             @"sendBreadcrumbForControlNotification: v24@0:8@16",
             @"sendBreadcrumbForMenuItemNotification: v24@0:8@16",
             @"sendBreadcrumbForNotification: v24@0:8@16",
@@ -107,9 +109,11 @@
             @"setNotifier: v24@0:8@16",
             @"setObserver: v24@0:8@?16",
             @"setPluginClient: v24@0:8@16",
+            @"setReadyForInternalCalls: v20@0:8B16",
+            @"setReadyForInternalCalls: v20@0:8c16",
             @"setSessionTracker: v24@0:8@16",
-            @"setStarted: v20@0:8B16",
-            @"setStarted: v20@0:8c16",
+            @"setIsStarted: v20@0:8B16",
+            @"setIsStarted: v20@0:8c16",
             @"setState: v24@0:8@16",
             @"setStateEventBlocks: v24@0:8@16",
             @"setStateMetadataFromLastLaunch: v24@0:8@16",
@@ -130,8 +134,8 @@
     // the following methods are implemented on Bugsnag but do not need to
     // be mirrored on BugsnagClient
     self.bugsnagMethodsNotRequiredOnClient = [NSSet setWithArray:@[
-            @"bugsnagStarted B16@0:8",
-            @"bugsnagStarted c16@0:8",
+            @"bugsnagReadyForInternalCalls B16@0:8",
+            @"bugsnagReadyForInternalCalls c16@0:8",
             @"client @16@0:8",
             @"getContext @16@0:8",
             @"purge v16@0:8",